requirements.txt
absl-py==2.1.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.3.2
altair==6.0.0
annotated-types==0.7.0
anyio==4.7.0
appdirs==1.4.4
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.4
async-timeout==5.0.1
attrs==25.4.0
audioread==3.0.1
autograd==1.7.0
autoray==0.7.0
babel==2.16.0
bayesian-optimization==3.1.0
beautifulsoup4==4.13.4
bibtexparser==1.4.3
biopython==1.86
blake3==1.0.8
bleach==6.2.0
blinker==1.9.0
boto3==1.42.16
botocore==1.42.16
cachetools==5.5.1
cattrs==25.3.0
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
chembl-webresource-client==0.10.9
click==8.1.8
comm==0.2.2
contourpy==1.3.1
cssselect2==0.8.0
cycler==0.12.1
datasets==3.4.0
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
diastatic-malt==2.15.2
diffusers==0.35.1
dill==0.3.8
diskcache==5.6.3
duckdb==1.4.3
easydict==1.13
easyocr==1.7.2
emmet-core==0.86.0rc1
# Editable install with no version control (eraslang==1.0)
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.1.0
faiss-cpu==1.10.0
fastapi==0.116.1
fastjsonschema==2.21.1
ffmpy==0.6.1
filelock==3.20.1
Flask==3.1.0
flask-cors==6.0.0
flatbuffers==24.12.23
fluidsynth==0.2
fonttools==4.60.2
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.2.0
gast==0.6.0
gevent==24.2.1
gitdb==4.0.12
GitPython==3.1.45
google-pasta==0.2.0
gradio==5.50.0
gradio_client==1.14.0
gradio_molecule3d==0.0.7
greenlet==3.2.4
groovy==0.1.2
grpcio==1.68.1
gunicorn==23.0.0
h11==0.16.0
h5py==3.12.1
hf-xet==1.1.10
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.35.1
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
ipykernel==6.29.5
ipython==8.31.0
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
json5==0.10.0
jsonpickle==4.1.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kagglehub==0.3.6
keras==3.7.0
keras-hub==0.18.1
keras-tuner==1.4.7
kiwisolver==1.4.8
kt-legacy==1.0.5
latexcodec==3.0.1
lazy_loader==0.4
libclang==18.1.1
librosa==0.11.0
llama_cpp_python==0.2.90
llvmlite==0.45.0
lyricsgenius==3.0.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mido==1.3.3
mistune==3.1.0
ml-dtypes==0.3.2
mlx==0.29.4
mlx-graphs==0.0.8
mlx-metal==0.29.2
monty==2025.3.3
more-itertools==10.8.0
mp-api==0.45.13
mpmath==1.3.0
msgpack==1.1.1
multidict==6.1.0
multiprocess==0.70.16
music21==9.7.1
namex==0.0.8
narwhals==2.14.0
nbclient==0.10.2
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
nltk==3.9.2
notebook==7.3.2
notebook_shim==0.2.4
numba==0.62.0
numpy==1.26.3
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opensimplex==0.4.5.1
opt_einsum==3.4.0
optree==0.13.1
orjson==3.11.3
outcome==1.3.0.post0
overrides==7.7.0
packaging==25.0
palettable==3.3.3
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
patsy==1.0.1
prefect==3.0.0
PennyLane==0.40.0
PennyLane_Lightning==0.40.0
pexpect==4.9.0
pillow==11.2.1
playsound==1.3.0
playwright==1.54.0
plotly==6.5.0
pooch==1.8.2
pretty_midi==0.2.10
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.3.0
protobuf==4.25.6
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pybtex==0.25.1
pyclipper==1.3.0.post6
pydantic==2.11.7
pydantic-settings==2.12.0
pydantic_core==2.33.2
pydeck==0.9.1
pydub==0.25.1
pydyf==0.11.0
pyee==13.0.0
pyfluidsynth==1.3.4
pygame==2.6.1
Pygments==2.18.0
pymatgen==2025.10.7
pymatgen-io-validation==0.1.2
PyOpenGL==3.1.10
PyOpenGL-accelerate==3.1.10
pyparsing==3.2.0
pyphen==0.17.2
PyQt6==6.9.1
PyQt6-Qt6==6.9.2
PyQt6_sip==13.10.2
pyqtgraph==0.13.7
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.2.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.2.0
rdkit==2025.9.1
redis==7.1.0
referencing==0.35.1
regex==2024.11.6
requests==2.31.0
requests-cache==1.2.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.22.3
ruff==0.13.2
rustworkx==0.16.0
s3transfer==0.16.0
safehttpx==0.1.7
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.0
scipy==1.15.2
scipy-openblas32==0.3.29.0.0
selenium==4.32.0
selfies==2.2.0
semantic-version==2.10.0
Send2Trash==1.8.3
shap==0.45.0
shapely==2.1.0
shellingham==1.5.4
six==1.17.0
smart_open==7.5.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
soupsieve==2.7
soxr==1.0.0
spglib==2.6.0
spotipy==2.25.2
stack-data==0.6.3
starlette==0.49.1
statsmodels==0.14.5
streamlit==1.52.2
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-macos==2.16.2
tensorflow-metal==1.2.0
termcolor==2.5.0
terminado==0.18.1
textstat==0.7.12
threadpoolctl==3.5.0
tifffile==2025.3.30
timm==1.0.20
tinycss2==1.4.0
tinyhtml5==2.0.0
tinytex==0.2.0
tokenizers==0.21.1
toml==0.10.2
tomli==2.2.1
tomlkit==0.13.2
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.53.0
trio==0.30.0
trio-websocket==0.12.2
typer==0.19.2
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.3
umap-learn==0.5.5
uncertainties==3.2.3
vaderSentiment==3.3.2
uri-template==1.3.0
url-normalize==2.2.1
urllib3==2.6.3
uvicorn==0.35.0
uvloop==0.21.0
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.13
weasyprint==66.0
webcolors==24.11.1
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.4
wget==3.2
wrapt==1.17.0
wsproto==1.2.0
xgboost==3.1.2
xxhash==3.5.0
yarl==1.18.3
zipp==3.23.0
zope.event==5.1.1
zope.interface==7.2
zopfli==0.2.3.post1
zstandard==0.22.0
